web/requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
gunicorn==20.0.4
Keras==2.4.3
Keras-Preprocessing==1.1.2
nltk==3.5
pytz==2020.1
scikit-learn==0.23.2
sqlparse==0.3.1
tensorflow==2.3.0
